Struct lmdb_rs::core::EnvBuilder

pub struct EnvBuilder {
    // some fields omitted
}

Constructs environment with settigs which couldn't be changed after opening. By default it tries to create corresponding dir if it doesn't exist, use autocreate_dir() to override that behavior

Methods

impl EnvBuilder

fn new() -> EnvBuilder

fn flags(self, flags: EnvCreateFlags) -> EnvBuilder

Sets environment flags

fn max_readers(self, max_readers: usize) -> EnvBuilder

Sets max concurrent readers operating on environment

fn max_dbs(self, max_dbs: usize) -> EnvBuilder

Set max number of databases

fn map_size(self, map_size: u64) -> EnvBuilder

Sets max environment size, i.e. size in memory/disk of all data

fn autocreate_dir(self, autocreate_dir: bool) -> EnvBuilder

Sets whetever lmdb-rs should try to autocreate dir with default permissions on opening (default is true)

fn open(self, path: &Path, perms: u32) -> MdbResult<Environment>

Opens environment in specified path