Struct lmdb_rs::core::ReadonlyTransaction

pub struct ReadonlyTransaction<'a> { /* fields omitted */ }

Methods

impl<'a> ReadonlyTransaction<'a>

fn new_ro_child(&self) -> MdbResult<ReadonlyTransaction>

fn abort(&mut self)

Aborts transaction. But readonly transaction could be reused later by calling renew

fn reset(&mut self)

Resets read only transaction, handle is kept. Must be followed by call to renew

fn renew(&mut self) -> MdbResult<()>

Acquires a new reader lock after transaction abort or reset

fn bind(&self, db_handle: &DbHandle) -> Database