Trait lmdb_rs::core::IterateCursor

pub trait IterateCursor {
    fn init_cursor<'a, 'b: 'a>(&'a self, cursor: &mut Cursor<'b>) -> bool;
    fn move_to_next<'iter, 'cursor: 'iter>(&'iter self,
                                           cursor: &'cursor mut Cursor<'cursor>)
                                           -> bool;

    fn get_size_hint(&self, _cursor: &Cursor) -> (usize, Option<usize>) { ... }
}

Allows the cration of custom cursor iteration behaviours.

Required Methods

fn init_cursor<'a, 'b: 'a>(&'a self, cursor: &mut Cursor<'b>) -> bool

Returns true if initialization successful, for example that the key exists.

fn move_to_next<'iter, 'cursor: 'iter>(&'iter self,
cursor: &'cursor mut Cursor<'cursor>)
-> bool

Returns true if there is still data and iterator is in correct range

Provided Methods

fn get_size_hint(&self, _cursor: &Cursor) -> (usize, Option<usize>)

Returns size hint considering current state of cursor