Trait lmdb_rs::traits::ToMdbValue

pub trait ToMdbValue {
    fn to_mdb_value<'a>(&'a self) -> MdbValue<'a>;
}

ToMdbValue is supposed to convert a value to a memory slice which lmdb uses to prevent multiple copying data multiple times. May be unsafe.

Required Methods

fn to_mdb_value<'a>(&'a self) -> MdbValue<'a>