Trait lmdb_rs::traits::FromMdbValue

pub trait FromMdbValue {
    fn from_mdb_value(value: &MdbValue) -> Self;
}

FromMdbValue is supposed to reconstruct a value from memory slice. It allows to use zero copy where it is required.

Required Methods

fn from_mdb_value(value: &MdbValue) -> Self